update: clean up logger and fix count bytes sent
