+++ internal/monitor/monitor.go
@@ -62,8 +62,8 @@ func (m *Monitor) Start(ctx context.Context) error {
 	m.running = true
 	m.mu.Unlock()
 
-	m.logger.Infof("\n=== Starting Event Monitor ===\n")
-	m.logger.Infof("Initializing monitor for all registered contracts...\n")
+	m.logger.Infof("\n=== Starting Event Monitor ===")
+	m.logger.Infof("Initializing monitor for all registered contracts...")
 
 	// Get all contracts from the database
 	contracts, err := m.db.Contract.FindMany().Exec(ctx)
@@ -74,13 +74,13 @@ func (m *Monitor) Start(ctx context.Context) error {
 		return fmt.Errorf("failed to get contracts: %v", err)
 	}
 
-	m.logger.Infof("Found %d contracts to monitor\n", len(contracts))
+	m.logger.Infof("Found %d contracts to monitor", len(contracts))
 
 	for chainID, client := range m.ethClients {
 		block, err := client.GetLatestBlock()
 		if err == nil {
 			m.lastBlocks[chainID] = block
-			m.logger.Infof("Initial block for chain %d: %d\n", chainID, block)
+			m.logger.Infof("Initial block for chain %d: %d", chainID, block)
 		}
 	}
 
@@ -98,7 +98,7 @@ func (m *Monitor) Start(ctx context.Context) error {
 
 		// Start a goroutine for each contract
 		go func(c interface{}) {
-			m.logger.Infof("Starting independent monitor for contract %s\n", c.(db.ContractModel).ID)
+			m.logger.Infof("Starting independent monitor for contract %s", c.(db.ContractModel).ID)
 			m.monitorContract(contractCtx, c)
 		}(contract)
 	}
@@ -109,11 +109,11 @@ func (m *Monitor) Start(ctx context.Context) error {
 	// Start a goroutine to periodically check for new contracts
 	go m.checkForNewContracts(ctx)
 
-	m.logger.Infof("Monitor started successfully. Each contract will be monitored independently.\n")
-	m.logger.Infof("Contracts being backfilled will be monitored concurrently with backfill.\n")
-	m.logger.Infof("Will log current blocks every 30 seconds.\n")
-	m.logger.Infof("Will check for new contracts every 10 seconds.\n")
-	m.logger.Infof("=== Event Monitor Ready ===\n\n")
+	m.logger.Infof("Monitor started successfully. Each contract will be monitored independently.")
+	m.logger.Infof("Contracts being backfilled will be monitored concurrently with backfill.")
+	m.logger.Infof("Will log current blocks every 30 seconds.")
+	m.logger.Infof("Will check for new contracts every 10 seconds.")
+	m.logger.Infof("=== Event Monitor Ready ===")
 
 	return nil
 }
@@ -137,7 +137,7 @@ func (m *Monitor) logCurrentBlock(ctx context.Context) {
 			for chainID, client := range m.ethClients {
 				block, err := client.GetLatestBlock()
 				if err == nil {
-					m.logger.Infof("Current block on chain %d: %d\n", chainID, block)
+					m.logger.Infof("Current block on chain %d: %d", chainID, block)
 				}
 			}
 			m.mu.RUnlock()
@@ -163,7 +163,7 @@ func (m *Monitor) Stop() {
 	m.backfilling = make(map[string]bool)
 	m.readyForMonitoring = make(map[string]bool)
 	m.running = false
-	m.logger.Infof("Stopped event monitor\n")
+	m.logger.Infof("Stopped event monitor")
 }
 
 func (m *Monitor) monitorContract(ctx context.Context, contract interface{}) {
@@ -188,15 +188,15 @@ func (m *Monitor) monitorContract(ctx context.Context, contract interface{}) {
 		id = contractValue.FieldByName("ID").String()
 	}
 
-	m.logger.Infof("\n=== Starting Monitor for Contract ===\n")
-	m.logger.Infof("Chain ID: %d\n", chainID)
-	m.logger.Infof("Address: %s\n", address)
-	m.logger.Infof("Event Signature: %s\n", eventSignature)
-	m.logger.Infof("Contract ID: %s\n", id)
+	m.logger.Infof("\n=== Starting Monitor for Contract ===")
+	m.logger.Infof("Chain ID: %d", chainID)
+	m.logger.Infof("Address: %s", address)
+	m.logger.Infof("Event Signature: %s", eventSignature)
+	m.logger.Infof("Contract ID: %s", id)
 
 	client, ok := m.ethClients[int(chainID)]
 	if !ok {
-		m.logger.Infof("ERROR: No Ethereum client found for chain ID %d\n", chainID)
+		m.logger.Infof("ERROR: No Ethereum client found for chain ID %d", chainID)
 		return
 	}
 
@@ -204,19 +204,19 @@ func (m *Monitor) monitorContract(ctx context.Context, contract interface{}) {
 	if eventABI != "" {
 		err := client.RegisterEventABI(eventSignature, eventABI)
 		if err != nil {
-			m.logger.Infof("WARNING: Failed to register event ABI: %v\n", err)
+			m.logger.Infof("WARNING: Failed to register event ABI: %v", err)
 		} else {
-			m.logger.Infof("Successfully registered event ABI for signature %s\n", eventSignature)
+			m.logger.Infof("Successfully registered event ABI for signature %s", eventSignature)
 		}
 	}
 
 	latestBlock, err := client.GetLatestBlock()
 	if err != nil {
-		m.logger.Infof("ERROR: Failed to get latest block: %v\n", err)
+		m.logger.Infof("ERROR: Failed to get latest block: %v", err)
 		return
 	}
 
-	m.logger.Infof("Starting to monitor from block %d\n", latestBlock)
+	m.logger.Infof("Starting to monitor from block %d", latestBlock)
 
 	m.mu.Lock()
 	if _, exists := m.lastBlocks[int(chainID)]; !exists {
@@ -225,34 +225,34 @@ func (m *Monitor) monitorContract(ctx context.Context, contract interface{}) {
 	lastProcessedBlock := m.lastBlocks[int(chainID)]
 	m.mu.Unlock()
 
-	m.logger.Infof("Last processed block: %d\n", lastProcessedBlock)
-	m.logger.Infof("=== Monitor Initialized ===\n\n")
+	m.logger.Infof("Last processed block: %d", lastProcessedBlock)
+	m.logger.Infof("=== Monitor Initialized ===")
 
 	pollTicker := time.NewTicker(1 * time.Second)
 	defer pollTicker.Stop()
 
 	lastBackfillStatus := m.IsContractBackfilling(id)
-	m.logger.Infof("Initial backfill status for contract %s: %v\n", id, lastBackfillStatus)
+	m.logger.Infof("Initial backfill status for contract %s: %v", id, lastBackfillStatus)
 
 	for {
 		select {
 		case <-ctx.Done():
-			m.logger.Infof("Stopping monitor for contract %s\n", address)
+			m.logger.Infof("Stopping monitor for contract %s", address)
 			return
 		case <-pollTicker.C:
 			currentBackfillStatus := m.IsContractBackfilling(id)
 			if currentBackfillStatus != lastBackfillStatus {
 				if currentBackfillStatus {
-					m.logger.Infof("Contract %s is now being backfilled\n", id)
+					m.logger.Infof("Contract %s is now being backfilled", id)
 				} else {
-					m.logger.Infof("Contract %s is no longer being backfilled\n", id)
+					m.logger.Infof("Contract %s is no longer being backfilled", id)
 				}
 				lastBackfillStatus = currentBackfillStatus
 			}
 
 			currentBlock, err := client.GetLatestBlock()
 			if err != nil {
-				m.logger.Infof("ERROR: Failed to get latest block: %v\n", err)
+				m.logger.Infof("ERROR: Failed to get latest block: %v", err)
 				continue
 			}
 
@@ -261,9 +261,9 @@ func (m *Monitor) monitorContract(ctx context.Context, contract interface{}) {
 			m.mu.Unlock()
 
 			if currentBlock > lastProcessedBlock {
-				// m.logger.Infof()("\n=== Processing New Blocks ===\n")
-				// m.logger.Infof()("Contract: %s\n", address)
-				// m.logger.Infof()("Processing blocks %d to %d\n", lastProcessedBlock+1, currentBlock)
+				// m.logger.Infof()("\n=== Processing New Blocks ===")
+				// m.logger.Infof()("Contract: %s", address)
+				// m.logger.Infof()("Processing blocks %d to %d", lastProcessedBlock+1, currentBlock)
 
 				batchSize := uint64(5)
 				for fromBlock := lastProcessedBlock + 1; fromBlock <= currentBlock; fromBlock += batchSize {
@@ -272,7 +272,7 @@ func (m *Monitor) monitorContract(ctx context.Context, contract interface{}) {
 						toBlock = currentBlock
 					}
 
-					m.logger.Infof("Fetching logs for blocks %d to %d\n", fromBlock, toBlock)
+					m.logger.Infof("Fetching logs for blocks %d to %d", fromBlock, toBlock)
 
 					// Get logs for the block range
 					logs, err := client.FilterLogs(
@@ -285,11 +285,11 @@ func (m *Monitor) monitorContract(ctx context.Context, contract interface{}) {
 					)
 
 					if err != nil {
-						m.logger.Infof("ERROR: Failed to fetch logs: %v\n", err)
+						m.logger.Infof("ERROR: Failed to fetch logs: %v", err)
 						continue
 					}
 
-					m.logger.Infof("Found %d logs for blocks %d to %d\n", len(logs), fromBlock, toBlock)
+					m.logger.Infof("Found %d logs for blocks %d to %d", len(logs), fromBlock, toBlock)
 
 					m.ProcessLogsEvent(ctx, logs, client, eventSignature, chainID, address)
 
@@ -303,7 +303,7 @@ func (m *Monitor) monitorContract(ctx context.Context, contract interface{}) {
 							ChainId:         int(chainID),
 						})
 						if err != nil {
-							m.logger.Infof("ERROR: Failed to send transaction: %v\n", err)
+							m.logger.Infof("ERROR: Failed to send transaction: %v", err)
 						}
 					}
 
@@ -313,11 +313,11 @@ func (m *Monitor) monitorContract(ctx context.Context, contract interface{}) {
 					lastProcessedBlock = toBlock
 					m.mu.Unlock()
 
-					m.logger.Infof("Updated last processed block to %d\n", lastProcessedBlock)
-					m.logger.Infof("=== Finished Processing Blocks ===\n\n")
+					m.logger.Infof("Updated last processed block to %d", lastProcessedBlock)
+					m.logger.Infof("=== Finished Processing Blocks ===")
 				}
 			} else {
-				m.logger.Infof("No new blocks to process. Current: %d, Last: %d\n", currentBlock, lastProcessedBlock)
+				m.logger.Infof("No new blocks to process. Current: %d, Last: %d", currentBlock, lastProcessedBlock)
 			}
 		}
 	}
@@ -327,13 +327,13 @@ func (m *Monitor) ProcessLogsEvent(ctx context.Context, logs []ethereum.Log, cli
 	for _, eventLog := range logs {
 		decodedData, err := client.GetDecoder().DecodeEvent(eventSignature, eventLog.Data, eventLog.Topics)
 		if err != nil {
-			m.logger.Infof("ERROR: Failed to decode event data: %v\n", err)
+			m.logger.Infof("ERROR: Failed to decode event data: %v", err)
 			decodedData = fmt.Sprintf("{\"raw\": \"%s\"}", common.Bytes2Hex(eventLog.Data))
 		}
 
 		err = m.processEvent(ctx, int(chainID), address, eventSignature, eventLog, decodedData)
 		if err != nil {
-			m.logger.Infof("ERROR: Failed to process event: %v\n", err)
+			m.logger.Infof("ERROR: Failed to process event: %v", err)
 			continue
 		}
 	}
@@ -411,7 +411,7 @@ func (m *Monitor) MarkContractBackfilling(contractID string) {
 
 	m.readyForMonitoring[contractID] = true
 
-	m.logger.Infof("Marked contract %s as being backfilled\n", contractID)
+	m.logger.Infof("Marked contract %s as being backfilled", contractID)
 }
 
 func (m *Monitor) MarkContractBackfillComplete(contractID string) {
@@ -422,7 +422,7 @@ func (m *Monitor) MarkContractBackfillComplete(contractID string) {
 
 	m.readyForMonitoring[contractID] = true
 
-	m.logger.Infof("Marked contract %s as having completed backfill\n", contractID)
+	m.logger.Infof("Marked contract %s as having completed backfill", contractID)
 }
 
 func (m *Monitor) IsContractBackfilling(contractID string) bool {
@@ -459,8 +459,8 @@ func (m *Monitor) RegisterContract(ctx context.Context, contract interface{}) er
 		id = contractValue.FieldByName("ID").String()
 	}
 
-	m.logger.Infof("\n=== Registering New Contract with Monitor ===\n")
-	m.logger.Infof("Contract ID: %s\n", id)
+	m.logger.Infof("\n=== Registering New Contract with Monitor ===")
+	m.logger.Infof("Contract ID: %s", id)
 
 	m.mu.Lock()
 	m.readyForMonitoring[id] = true
@@ -472,12 +472,12 @@ func (m *Monitor) RegisterContract(ctx context.Context, contract interface{}) er
 	m.mu.Unlock()
 
 	go func(c interface{}, id string) {
-		m.logger.Infof("Starting independent monitor for newly registered contract %s\n", id)
+		m.logger.Infof("Starting independent monitor for newly registered contract %s", id)
 		m.monitorContract(contractCtx, c)
 	}(contract, id)
 
-	m.logger.Infof("Successfully registered contract %s with monitor\n", id)
-	m.logger.Infof("=== Contract Registration Complete ===\n\n")
+	m.logger.Infof("Successfully registered contract %s with monitor", id)
+	m.logger.Infof("=== Contract Registration Complete ===")
 
 	return nil
 }
@@ -500,7 +500,7 @@ func (m *Monitor) checkForNewContracts(ctx context.Context) {
 
 			contracts, err := m.db.Contract.FindMany().Exec(ctx)
 			if err != nil {
-				m.logger.Infof("Error getting contracts: %v\n", err)
+				m.logger.Infof("Error getting contracts: %v", err)
 				continue
 			}
 
@@ -515,9 +515,9 @@ func (m *Monitor) checkForNewContracts(ctx context.Context) {
 					continue
 				}
 
-				m.logger.Infof("Found new contract %s, registering with monitor\n", contractID)
+				m.logger.Infof("Found new contract %s, registering with monitor", contractID)
 				if err := m.RegisterContract(ctx, contract); err != nil {
-					m.logger.Infof("Error registering contract %s: %v\n", contractID, err)
+					m.logger.Infof("Error registering contract %s: %v", contractID, err)
 				}
 			}
 		}
@@ -594,7 +594,7 @@ func (m *Monitor) SendTx(ctx context.Context, filter types.FilterLogsParams) err
 		return err
 	}
 
-	m.logger.Infof("Transaction ID: %s\n", response["txHash"])
+	m.logger.Infof("Transaction ID: %s", response["txHash"])
 
 	return nil
 }
